Feature/ble datatypes

Approved-by: Alexander Mykhailiutenko

+++ shared/src/commonMain/kotlin/com/gorman/archimed/states/bluetooth/ExperimentOnlineData.kt
@@ -3,6 +3,6 @@ package com.gorman.archimed.states.bluetooth
 import com.gorman.bluetooth.constants.SensorType
 
 data class ExperimentOnlineData(
-    val currentSample: Int = 0,
+    val currentSample: Short = 0,
     val sensorsData: Map<SensorType, Double> = mapOf()
 )
